refactor: move model to global
