+++ src/main/java/com/avalon/avalonchat/core/login/repository/RefreshTokenRepository.java
@@ -2,7 +2,7 @@ package com.avalon.avalonchat.core.login.repository;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.avalon.avalonchat.core.model.RefreshToken;
+import com.avalon.avalonchat.global.model.RefreshToken;
 
 public interface RefreshTokenRepository extends CrudRepository<RefreshToken, String> {
 }
