+++ src/main/java/com/avalon/avalonchat/core/profile/domain/ProfileImage.java
@@ -9,7 +9,7 @@ import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
-import com.avalon.avalonchat.core.model.BaseAuditingEntity;
+import com.avalon.avalonchat.global.model.BaseAuditingEntity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
