+++ src/main/java/com/avalon/avalonchat/global/model/BaseAuditingEntity.java
@@ -1,4 +1,4 @@
-package com.avalon.avalonchat.core.model;
+package com.avalon.avalonchat.global.model;
 
 import java.time.LocalDateTime;
