+++ src/test/java/com/avalon/avalonchat/core/login/repository/RefreshTokenRepositoryTest.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.avalon.avalonchat.core.model.RefreshToken;
+import com.avalon.avalonchat.global.model.RefreshToken;
 import com.avalon.avalonchat.testsupport.base.BaseTestContainerTest;
 
 @SpringBootTest
